Add files via upload
In this projects discussed about Area Plots, Histograms, and Bar Charts.
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
After completing this lab I will be able to:

Create and customize
1.Area plot
2.Histogram
3.Bar charts on a dataset

-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Area Plots
In the last module, we created a line plot that visualized the top 5 countries that contribued the most immigrants to Canada from 1980 to 2013. With a little modification to the code, we can visualize this plot as a cumulative plot, also knows as a Stacked Line Plot or Area plot.

-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Histograms
A histogram is a way of representing the frequency distribution of numeric dataset. The way it works is it partitions the x-axis into bins, assigns each data point in our dataset to a bin, and then counts the number of data points that have been assigned to each bin. So the y-axis is the frequency or the number of data points in each bin. Note that we can change the bin size and usually one needs to tweak it so that the distribution is displayed nicely.


-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Bar Charts (Dataframe) 
A bar plot is a way of representing data where the length of the bars represents the magnitude/size of the feature/variable. Bar graphs usually represent numerical and categorical variables grouped in intervals.

To create a bar plot, we can pass one of two arguments via kind parameter in plot():

kind=bar creates a vertical bar plot
kind=barh creates a horizontal bar plot
Vertical bar plot

In vertical bar graphs, the x-axis is used for labelling, and the length of bars on the y-axis corresponds to the magnitude of the variable being measured. Vertical bar graphs are particularly useful in analyzing time series data. One disadvantage is that they lack space for text labelling at the foot of each bar.